requirements.txt
langchain
torch
accelerate
bitsandbytes
transformers
sentence_transformers
faiss_cpu
chainlit
ctransformers
textract
tika
autoawq